tickets/src/events/listeners/__test__/order-created-listener.test.ts
import mongoose from 'mongoose';
import { OrderCreatedEvent, OrderStatus } from '@ticketo/common';
import { OrderCreatedListener } from '../order-created-listener';
import { natsWrapper } from '../../../nats-wrapper';
import { Ticket } from '../../../models/ticket';

const setup = async () => {
  //Create an instance of listener
  const listener = new OrderCreatedListener(natsWrapper.client);

  //create a ticket
  const ticket = Ticket.build({
    title: 'Concert',
    price: 100,
    userId: mongoose.Types.ObjectId().toHexString(),
  });

  await ticket.save();

  //Create a fake data event and fake msg instance
  const data: OrderCreatedEvent['data'] = {
    id: mongoose.Types.ObjectId().toHexString(),
    version: 0,
    status: OrderStatus.Created,
    userId: ticket.userId,
    expiresAt: new Date().toISOString(),
    ticket: {
      id: ticket.id,
      price: ticket.price,
    },
  };

  // @ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, ticket, data, msg };
};

it('updates the orderId of a ticket from the data object from the event', async () => {
  const { listener, ticket, data, msg } = await setup();
  await listener.onMessage(data, msg);

  const updatedTicket = await Ticket.findById(ticket.id);
  expect(updatedTicket!.orderId).toEqual(data.id);
});

it('calls the ack function in msg', async () => {
  const { listener, ticket, data, msg } = await setup();
  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});

it('publishes after the ticket is updated with the orderId', async () => {
  const { listener, ticket, data, msg } = await setup();
  await listener.onMessage(data, msg);
  expect(natsWrapper.client.publish).toHaveBeenCalled();

  const dataFromThePublishCall = JSON.parse(
    (natsWrapper.client.publish as jest.Mock).mock.calls[0][1]
  );

  expect(data.id).toEqual(dataFromThePublishCall.orderId);
});
